Move validation of CloudFormation templates to its own module
